Remove the concept of Organization
This will convert the project from multi-tenant to single tenant software. To support multiple tenants, simply deploy separate instance of it with separate databases.

The rational behind this change is to reduce code complexity. From now on, the software is supposed to be for one tenant (org) and everything by default belongs to that tenant.