timestamp --> timestamptz
This is done for DateTime fields whose content is directly saved for data received externally (e.g. transaction time, notify event time). In such cases, it is important to store the timezone information.

For fields that are set internally only (e.g. createdAt, updatedAt), timezone isn't required since we know that it will always be in the server's timezone.